requirements.txt
# Core dependencies
pyyaml>=6.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0